# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ParchiFlow - Universal SME Ledger Intelligence & Collection Automation OS
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports/party-statement/{partyId}:
    get:
      operationId: getPartyStatement
      tags: [reports]
      summary: Get party statement
      parameters:
        - name: partyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Party statement
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PartyStatement"
components:
  schemas:
...
    Party:
      type: object
      properties:
        id:
          type: integer
        businessId:
          type: integer
        name:
          type: string
        normalizedName:
          type: string
        mobile:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        gstin:
          type: string
          nullable: true
        type:
          type: string
        address:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        currentBalance:
          type: number
        balanceType:
          type: string
        openingBalance:
          type: number
        openingBalanceType:
          type: string
        riskScore:
          type: number
          nullable: true
        sourceType:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, businessId, name, normalizedName, type, currentBalance, balanceType, openingBalance, openingBalanceType, createdAt]
...
    LedgerEntry:
      type: object
      properties:
        id:
          type: integer
        businessId:
          type: integer
        partyId:
          type: integer
          nullable: true
        partyName:
          type: string
          nullable: true
        entryType:
          type: string
        voucherNumber:
          type: string
          nullable: true
        invoiceNumber:
          type: string
          nullable: true
        amount:
          type: number
        debitAmount:
          type: number
          nullable: true
        creditAmount:
          type: number
          nullable: true
        entryDate:
          type: string
          format: date-time
        dueDate:
          type: string
          format: date-time
          nullable: true
        narration:
          type: string
          nullable: true
        status:
          type: string
        reconciliationStatus:
          type: string
        sourceType:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, businessId, entryType, amount, entryDate, status, reconciliationStatus, createdAt]
...
    PartyStatement:
      type: object
      properties:
        party:
          $ref: "#/components/schemas/Party"
        openingBalance:
          type: number
        openingBalanceType:
          type: string
        closingBalance:
          type: number
        closingBalanceType:
          type: string
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/LedgerEntry"
        totalDebits:
          type: number
        totalCredits:
          type: number
      required: [party, openingBalance, openingBalanceType, closingBalance, closingBalanceType, transactions, totalDebits, totalCredits]